refactor(asset_connector): add a mixin class
