+++ sekoia_automation/asset_connector/__init__.py
@@ -1,10 +1,12 @@
 from .async_connector import AsyncAssetConnector
 from .connector import AssetConnector
+from .mixin import AssetConnectorMixin
 from .utils import RATE_LIMIT_DEFAULT_WAIT, parse_retry_after
 
 __all__ = [
     "AssetConnector",
     "AsyncAssetConnector",
+    "AssetConnectorMixin",
     "RATE_LIMIT_DEFAULT_WAIT",
     "parse_retry_after",
 ]
